refactor: rename Handlers to Authenticators in namespaces
